unix-manager: add pcap-file command
This patch adds a 'pcap-file' commands which is available in the
new 'unix-socket' running mode.

When starting in 'unix-socket' mode, only a unix socket manager
is started. When it receives a 'pcap-file' command, the manager
start a 'pcap-file' running mode which does not really leave at
the end of file but simply exit. The manager is then able to start
a new running mode with a new file.