SMTP MIME Email Message decoder
This module parses email messages and decodes content bodies that are transmitted over the SMTP protocol.
Parsing is done on a line-by-line streaming basis to minimize memory resources.  Some protocol anomalies are
reported as flags in the data structure and file extraction is also available in a streaming fashion using
a callback function.